removed arithmetic.rs
