+++ crabml-core/src/backends/cpu/mod.rs
@@ -1,8 +1,6 @@
-pub mod arithmetic;
 pub mod buf;
 pub mod cpu_tensor;
 mod primitives;
-pub mod validate;
 
 pub use cpu_tensor::CpuTensor;
 pub use cpu_tensor::CpuTensorDeviceRef;
